Sort & format
Review by: rjrjr@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@7699 8db76d5a-ed1c-0410-87a9-c151d255dfc7

+++ bikeshed/src/com/google/gwt/sample/expenses/client/ValuesImpl.java
@@ -68,11 +68,11 @@ public final class ValuesImpl<T> extends JavaScriptObject implements Values<T> {
     this.propertyHolder = propertyHolder;
   }-*/;
 
-  private native int getInt(String name) /*-{
+  private native double getDouble(String name) /*-{
     return this[name];
   }-*/;
 
-  private native double getDouble(String name) /*-{
+  private native int getInt(String name) /*-{
     return this[name];
   }-*/;
